Move db helper
