+++ src/sync/index.js
@@ -7,7 +7,8 @@ const updateLocalTx = require('./updateLocalTx');
 const { getInstance } = require('../qclient');
 const { withdrawType } = require('../constants');
 const { getContractMetadata } = require('../config');
-const { db, DBHelper } = require('../db');
+const { db } = require('../db');
+const { DBHelper } = require('../db/db-helper');
 const { getLogger } = require('../utils/logger');
 const { publishSyncInfo } = require('../schema/subscriptions');
 const Topic = require('../models/topic');
