+++ src/sync/updateLocalTx.js
@@ -8,7 +8,8 @@ const bodhiToken = require('../api/bodhi_token');
 const eventFactory = require('../api/event_factory');
 const centralizedOracle = require('../api/centralized_oracle');
 const decentralizedOracle = require('../api/decentralized_oracle');
-const { db, DBHelper } = require('../db');
+const { db } = require('../db');
+const { DBHelper } = require('../db/db-helper');
 const { Config, getContractMetadata } = require('../config');
 const { txState } = require('../constants');
 const Utils = require('../utils');
